requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.7.0
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fastapi==0.80.0
filetype==1.2.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.263
langchain-experimental==0.0.8
langchainplus-sdk==0.0.20
langsmith==0.0.19
libmagic==1.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
platformdirs==3.10.0
playwright==1.37.0
pycodestyle==2.11.0
pydantic==1.10.8
pyee==9.0.4
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.19.1
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.9.0
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
yarl==1.9.2
